Add `std` feature and make it the default
This enables `serde_json` dependency, which is needed for deserializing
response messages. This will mean that client-side API will not be
available for `no_std` environments. This isn't as big an issue as it
would seem since typically on micrcontrolers, one would implement
services, not clients. Morever, this limitation will hopefully be lifted
once [`serde-json-core` supports complex enums][ce].

On the bright side, we can now make `serde-json-core` and `heapless`
optional dependencies, which will help `std` users avoid these
dependencies that they are unlikely to otherwise depend on.

This also adds a `embedded` feature, which will enable `heapless` and
`serde-json-core` dependencies. Users must enable either `std` or
`embedded` feature to use this crate.

[ce]: https://github.com/rust-embedded-community/serde-json-core/issues/94