✨ Require `alloc` feature of serde for embedded
Instead of documenting that user needs to enable that. This means that
`zarlink` is no longer no_alloc. 😥 While a bit sad, this is not a
complete deal-breaker for embedded still since typically an allocator is
easily pluggable. We'll still do our best to reduce allocations as much
as possible though as the heap on embedded targets is typically very
limited.