Flatten Result return type of Connection::receive_reply
We achieve this by adding a `Reply` variant in `Error` and a generic on
the `Error` type for it, defaulting to `&'static str`. The default will
never really be used in practice but it's helpful for the majority of
cases where the `Reply` variant will not be created or is even relevant.