🚑️ Fix a soundness issue in the internal MethodStream
There is no safe way to implement a lending stream/iterator in Rust
currnetly so the stream can not take ownership of the ReadConnection it
borrows from. Although it's internal type and we our use doesn't lead to
unsoundness (I think) since we don't keep the references to the buffer
inside the ReadConnection after it's dropped, we might accidentally
change the using code in the future and end up with very hard to debug
problems.