🧵 Service::handle now blocking
Otherwise the `Server::run` becomes `!Send`, making it impossible for
users to run it in spawned tasks in multi-threadded runtimes (e.g
`tokio::task::spawn`). Implementers can still handle methods
asynchronously by returning a stream. Although that means they can not
keep a reference to the server around in their async handling logic and
are limited to only be able to send owned types.

An alternative here would be to accept `Server::run` being `!Send` but
givne the widespread use of multi-threaded runtimes and (especially)
`tokio::task::spawn`, I have a strong feeling that it would make the API
very unattractive for the majority of the potential users.