Revert "🧵 Service::handle now blocking"
This reverts commit 2491fea18c6a853df93ca287ba3d8da9a01e18c3.

When implementing the async handling, I realized how limiting that is
w/o having the whole handler async. With the reverted change, a lot of
async code will be forced to use interior mutability and method handling
being non-async, would mean that either handlers will be limited to sync
locks (which isn't what you always want in async code) or having to spawn
tasks for all method handling. This especially sucks given that the whole
Server machinery is completely single-threaded itself.

Moreover, it looks very likely that the reason we can't just
`tokio::spawn` `Server::run` is not because the future is always `!Send`
but rather that the compiler is unable to determine that to be the case
because of known issue[1].

[1]: https://github.com/rust-lang/rust/issues/100013