📝 Update TODO
Drop the idea of reply stream not yielding `Reply` type. This would have
allowed us to make the developer experience better by not bothering with
Reply wrapping but the downside would have been that the client-side would
receive an extra reply, which is probably fine for most cases but not
always. Especially if you're just returning a oneshot stream.