✨ Add `idl` module
Add API to handle parsing and serialization of [Varlink Interface
Definition Language](https://varlink.org/Interface-Definition).

This module is only enabled when `idl` feature is enabled. The parsing
unfortunately requires allocation and hence need the `std` feature.
Fortunately parsing is not required on the server side, which is the
interesting part of embedded targets.