🔥 core: Drop `Debug` constraint on `Service::ReplyStream`
This constraint makes it impossible to return a `futures::BoxStream` from
the service method handlers and since we don't really need `Debug` impl
from the reply streams, we can just drop it.

Note however that we still have the Debug constaint on other types but I
think that wouldn't be a (at least major) problem.

Fixes #96.